Add python ssl bio for SSL/TLS (#29)
The standard library offers a first-class sans-io API to SSL/TLS inside
the ssl module. It uses the BIO interface of OpenSSL under the hood
which explicitly is about managing a SSL/TLS connection with buffers
instead of actual network reads/writes.

Even if it is part of the standard library and correctly documented, a
"python sans-io tls" query on search engines doesn't immediately yield
a result to the standard library and someone clueless (like the author
of this commit) can miss it.

A link is now provided inside the protocol list so it can be more
easily found.